+++ CHANGES.rst
@@ -264,8 +264,10 @@ API changes
 
   - WCS objects can now be initialized with an ImageHDU or
     PrimaryHDU object. [#4493]
-  - astropy.wcs now issues an INFO message when the header has SIP coefficients but 
+
+  - astropy.wcs now issues an INFO message when the header has SIP coefficients but
     "-SIP" is missing from CTYPE. [#4814]
+
 
 Bug fixes
 ^^^^^^^^^
@@ -360,8 +362,10 @@ Bug fixes
 
 - ``astropy.wcs``
 
-  - astropy.wcs.to_header removes "-SIP" from CTYPE when SIP coefficients 
-    are not written out, i.e. ``relax`` is either ``False`` or ``None``. [#4814]
+  - astropy.wcs.to_header removes "-SIP" from CTYPE when SIP coefficients
+    are not written out, i.e. ``relax`` is either ``False`` or ``None``.
+    astropy.wcs.to_header appends "-SIP" to CTYPE when SIP coefficients
+    are written out, i.e. ``relax=True``. [#4814]
 
 Other Changes and Additions
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^